# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend APIs for ETix",
        "version": "1.0.5",
        "title": "ETix API Reference",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "paccy7002@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/routes/{id}": {
            "get": {
                "tags": [
                    "route"
                ],
                "summary": "Find route by ID",
                "description": "Returns a single route",
                "operationId": "getRouteById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of transport route to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Route"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
...
        "/api/users/{email}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Find user by email",
                "description": "Returns a single user",
                "operationId": "getUserByEmail",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email of course to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid email supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"

        }
    },
    "definitions": {
...
        "Route": {
            "type": "object",
            "required": [
                "origin",
                "destination",
                "numberOfTickets",
                "transportCompany"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "origin": {
                    "type": "string",
                    "example": "NYABUGOGO"
                },
                "destination": {
                    "type": "string",
                    "example": "NYABIHU"
                },
                "distance": {
                    "type": "string",
                    "example": "150Km"
                },
                "duration": {
                    "type": "string",
                    "example": "3hr"
                },
                "numberOfTickets": {
                    "type": "number",
                    "example": "100"
                },
                "transportCompany": {
                    "type": "object",
                    "example": "VOLCANO"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "roles":{
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
...
}